fix(refactor): move calculateElapsedTime, loadActivePomodoro to appropiate modules, update imports
