+++ app/tests/unit/components/molecules/Counter.test.tsx
@@ -3,9 +3,9 @@ import { act, cleanup, fireEvent, render, screen } from '@testing-library/react'
 import testIds from '@data/testIds'
 import useCounter from '@hooks/useCounter'
 import Counter from '@molecules/Counter'
-import { calculateElapsedTime } from '@molecules/Counter/functions'
 import { pomodoroStore } from '@store/Pomodoro'
 import { resetStore } from '@tests/unit/helpers'
+import { calculateElapsedTime } from '@utils/time'
 
 const CounterWrapper = () => {
   useCounter()
